Move simple-tree out of feature-libraries
