+++ experimental/dds/tree2/src/feature-libraries/index.ts
@@ -65,6 +65,7 @@ export {
 	normalizeNewFieldContent,
 	NewFieldContent,
 	assertAllowedValue,
+	isFluidHandle,
 } from "./contextuallyTyped";
 
 export { ForestSummarizer } from "./forestSummarizer";
@@ -244,26 +245,9 @@ export {
 	Context,
 	TreeEvent,
 	EditableTreeEvents,
+	FlexTreeUnknownUnboxed,
+	onNextChange,
 } from "./flex-tree";
-
-export {
-	getProxyForField,
-	TreeObjectNodeFields,
-	TreeField,
-	TreeFieldInner,
-	TypedNode,
-	TreeNodeUnion,
-	TreeListNode,
-	TreeMapNode,
-	TreeObjectNode,
-	TreeRoot,
-	Tree,
-	TreeApi,
-	TreeNode,
-	TreeObjectFactory,
-	FactoryTreeSchema,
-	addFactory,
-} from "./simple-tree";
 
 export { treeSchemaFromStoredSchema } from "./storedToViewSchema";
