+++ experimental/dds/tree2/src/index.ts
@@ -218,27 +218,30 @@ export {
 	schemaIsObjectNode,
 	AllowedTypeSet,
 	SchemaBuilderOptions,
-	TreeListNode,
-	TreeObjectNodeFields,
-	TreeField,
-	TreeFieldInner,
-	TypedNode,
-	TreeNodeUnion,
-	TreeMapNode,
-	TreeObjectNode,
-	Tree,
-	TreeApi,
-	TreeNode,
 	FlexTreeTyped,
 	TreeEvent,
-	TreeObjectFactory,
 	SchemaCollection,
-	FactoryTreeSchema,
 	TreeCompressionStrategy,
 	treeSchemaFromStoredSchema,
 	encodeTreeSchema,
 	stackTreeFieldCursor,
 } from "./feature-libraries";
+
+export {
+	TreeObjectNodeFields,
+	TreeField,
+	TreeFieldInner,
+	TypedNode,
+	TreeNodeUnion,
+	TreeListNode,
+	TreeMapNode,
+	TreeObjectNode,
+	Tree,
+	TreeApi,
+	TreeNode,
+	TreeObjectFactory,
+	FactoryTreeSchema,
+} from "./simple-tree";
 
 export {
 	ISharedTree,
