+++ experimental/dds/tree2/src/test/simple-tree/primitives.spec.ts
@@ -4,8 +4,8 @@
  */
 
 import { strict as assert } from "assert";
-import { leaf, SchemaBuilder } from "../../../domains";
-import { treeViewWithContent } from "../../utils";
+import { leaf, SchemaBuilder } from "../../domains";
+import { treeViewWithContent } from "../utils";
 import { pretty } from "./utils";
 
 const _ = new SchemaBuilder({ scope: "test", libraries: [leaf.library] });
