+++ experimental/dds/tree2/src/test/simple-tree/rawObjectNode.spec.ts
@@ -9,15 +9,15 @@
 import { strict as assert } from "node:assert";
 
 import { validateAssertionError } from "@fluidframework/test-runtime-utils";
-import { leaf, SchemaBuilder } from "../../../domains";
-import { boxedIterator } from "../../../feature-libraries/flex-tree";
+import { leaf, SchemaBuilder } from "../../domains";
+import { boxedIterator } from "../../feature-libraries/flex-tree";
 import {
 	createRawObjectNode,
 	extractRawNodeContent,
 	rawObjectErrorMessage,
-} from "../../../feature-libraries/simple-tree/rawObjectNode";
-import { brand } from "../../../util";
-import { contextWithContentReadonly } from "../flex-tree/utils";
+} from "../../simple-tree/rawObjectNode";
+import { brand } from "../../util";
+import { contextWithContentReadonly } from "../feature-libraries/flex-tree/utils";
 
 describe("raw object nodes", () => {
 	function getRawObjectNode() {
